+++ src/core/client/extensions/Meta.ts
@@ -1,5 +1,6 @@
 import { Appearance } from '../../shared/interfaces/Appearance';
 import { Item } from '../../shared/interfaces/Item';
+import { Vehicle } from '../../shared/interfaces/Vehicle';
 
 export interface Meta {
     permissionLevel: number; // Used to determine the player's current permissionLevel as a player..
@@ -21,4 +22,7 @@ export interface Meta {
     inventory: Array<Array<Partial<Item>>>;
     equipment: Array<Partial<Item>>;
     toolbar: Array<Partial<Item>>;
+
+    // Vehicles
+    vehicles: Array<Vehicle>;
 }